+++ Changelog.md
@@ -10,6 +10,7 @@
 - Fixed N+1 query problem in StudentsController by eager loading user association (#7678)
 - Resque Host Authorization, removing env condition as this is for all environments (#7671)
 - Fixed ordering of assignments in the Assignment dropdown menu and Assignment index page (#7642)
+- Updated Section model associations with appropriate dependent options to handle cascade deletion while preventing deletion when students exist (#7681)
 - Prevent grade change in the grades table for a Marks Spreadsheet, when scrolling up or down with mouse or keys (#7680)
 
 ## [v2.8.1]